+++ xformers/ops/tiled_matmul.py
@@ -5,15 +5,11 @@
 
 
 import os
-from typing import List, Optional
+from typing import List, Tuple
 
 import torch
-import torch.multiprocessing.reductions
-from torch.utils._pytree import tree_flatten, tree_unflatten
-from typing_extensions import Annotated
 
 from .. import _is_triton_available
-from .common import Alias, make_pytorch_operator_for_dispatch_key
 
 
 # Copied over from the sequence parallel fused ops.
@@ -33,16 +29,10 @@ def _should_use_triton(device: torch.device, dtype: torch.dtype) -> bool:
     return True
 
 
-# We can't use make_pytorch_cuda_operator because PyTorch isn't able to inspect
-# Tensor[][] args to detect they contain CUDA args. Thus we need to register
-# this as a fallback implementation, so it gets invoked regardless of the args.
-# See: https://github.com/pytorch/pytorch/issues/113022
-@make_pytorch_operator_for_dispatch_key("")
-def tiled_matmul_fwd(
+def check_inputs(
     a: List[List[torch.Tensor]],
     b: List[List[torch.Tensor]],
-    out: Optional[List[List[Annotated[torch.Tensor, Alias("a", write=True)]]]] = None,
-) -> List[List[Annotated[torch.Tensor, Alias("a", write=True)]]]:
+) -> Tuple[List[int], List[int], List[int]]:
     assert len(a) >= 1 and len(a[0]) >= 1 and all(len(row) == len(a[0]) for row in a), (
         "the first operand must be a non-empty two-dimensional regular list of lists "
         "of tenors"
@@ -99,43 +89,58 @@ def tiled_matmul_fwd(
         )
     ks = aks
 
-    if out is not None:
-        assert (
-            len(out) >= 1
-            and len(out[0]) >= 1
-            and all(len(row) == len(out[0]) for row in out)
-        ), "out must be a non-empty two-dimensional regular list of lists of tenors"
-        assert len(out) == m_tiles
-        assert len(out[0]) == n_tiles
-        cms = [out[tile_m][0].shape[0] for tile_m in range(m_tiles)]
-        cns = [out[0][tile_n].shape[1] for tile_n in range(n_tiles)]
-        for tile_m in range(m_tiles):
-            for tile_n in range(n_tiles):
-                assert out[tile_m][tile_n].shape[0] == cms[tile_m], (
-                    f"the tensors on row {tile_m} of out must all have the same size "
-                    f"along the m dimension, got {cms[tile_m]} at position 0 and "
-                    f"{out[tile_m][tile_n].shape[0]} at position {tile_n}"
-                )
-                assert out[tile_m][tile_n].shape[1] == cns[tile_n], (
-                    f"the tensors on column {tile_n} of out must all have the same "
-                    f"size along the k dimension, got {cns[tile_n]} at position 0 and "
-                    f"{out[tile_m][tile_n].shape[1]} at position {tile_m}"
-                )
-        for tile_m in range(m_tiles):
-            assert cms[tile_m] == ms[tile_m], (
-                f"the tensors on row {tile_m} of out and those on row {tile_m} of the "
-                f"first operand must have the same size along the m dimension, got "
-                f"{cms[tile_m]} and {ms[tile_m]}"
-            )
+    return ms, ns, ks
+
+
+def check_output(out: List[List[torch.Tensor]], ms: List[int], ns: List[int]) -> None:
+    m_tiles, n_tiles = len(ms), len(ns)
+    assert (
+        len(out) >= 1
+        and len(out[0]) >= 1
+        and all(len(row) == len(out[0]) for row in out)
+    ), "out must be a non-empty two-dimensional regular list of lists of tenors"
+    assert len(out) == m_tiles
+    assert len(out[0]) == n_tiles
+    cms = [out[tile_m][0].shape[0] for tile_m in range(m_tiles)]
+    cns = [out[0][tile_n].shape[1] for tile_n in range(n_tiles)]
+    for tile_m in range(m_tiles):
         for tile_n in range(n_tiles):
-            assert cns[tile_n] == ns[tile_n], (
-                f"the tensors on column {tile_n} of out and those on column {tile_n} "
-                f"of the second operand must have the same size along the n dimension, "
-                f"got {cns[tile_n]} and {ns[tile_n]}"
+            assert out[tile_m][tile_n].shape[0] == cms[tile_m], (
+                f"the tensors on row {tile_m} of out must all have the same size "
+                f"along the m dimension, got {cms[tile_m]} at position 0 and "
+                f"{out[tile_m][tile_n].shape[0]} at position {tile_n}"
             )
-        c = out
-    else:
-        c = [[a[0][0].new_empty((m, n)) for n in ns] for m in ms]
+            assert out[tile_m][tile_n].shape[1] == cns[tile_n], (
+                f"the tensors on column {tile_n} of out must all have the same "
+                f"size along the k dimension, got {cns[tile_n]} at position 0 and "
+                f"{out[tile_m][tile_n].shape[1]} at position {tile_m}"
+            )
+    for tile_m in range(m_tiles):
+        assert cms[tile_m] == ms[tile_m], (
+            f"the tensors on row {tile_m} of out and those on row {tile_m} of the "
+            f"first operand must have the same size along the m dimension, got "
+            f"{cms[tile_m]} and {ms[tile_m]}"
+        )
+    for tile_n in range(n_tiles):
+        assert cns[tile_n] == ns[tile_n], (
+            f"the tensors on column {tile_n} of out and those on column {tile_n} "
+            f"of the second operand must have the same size along the n dimension, "
+            f"got {cns[tile_n]} and {ns[tile_n]}"
+        )
+
+
+# Using out= args in PyTorch is complicated, especially with custom_ops and
+# torch.compile (we need to declare our side-effects with mutates_args, which
+# then requires a functionalization step, ...). Thus this out= variant of the
+# operator is exposed as a PLAIN PYTHON function, and is not compilable nor
+# differentiable. It needs to be invoked from within a custom_op elsewhere.
+def tiled_matmul_out(
+    a: List[List[torch.Tensor]],
+    b: List[List[torch.Tensor]],
+    out: List[List[torch.Tensor]],
+) -> None:
+    ms, ns, ks = check_inputs(a, b)
+    check_output(out, ms, ns)
 
     # TODO We can try merging tiles that come from contiguous memory, using
     # stack_or_none, to further improve performance.
@@ -143,51 +148,119 @@ def tiled_matmul_fwd(
     # Because the Triton kernel is hardcoded for maximum three tiles.
     # Because, in turn, we aimed this at the fusion of wq/wk/wv.
     if (
-        m_tiles <= 3
-        and k_tiles <= 3
-        and n_tiles <= 3
+        len(ms) <= 3
+        and len(ks) <= 3
+        and len(ns) <= 3
         and _should_use_triton(a[0][0].device, a[0][0].dtype)
     ):
         from ._triton.tiled_matmul_kernels import _launch_triton_matmul
 
-        _launch_triton_matmul(a, b, c, ms, ns, ks)
+        _launch_triton_matmul(a, b, out, ms, ns, ks)
     else:
         for tile_m in range(len(ms)):
             for tile_n in range(len(ns)):
-                torch.mm(a[tile_m][0], b[0][tile_n], out=c[tile_m][tile_n])
+                torch.mm(a[tile_m][0], b[0][tile_n], out=out[tile_m][tile_n])
                 for tile_k in range(1, len(ks)):
-                    c[tile_m][tile_n].addmm_(a[tile_m][tile_k], b[tile_k][tile_n])
-
-    return c
+                    out[tile_m][tile_n].addmm_(a[tile_m][tile_k], b[tile_k][tile_n])
 
 
-def _transpose(x: List[List[torch.Tensor]]) -> List[List[torch.Tensor]]:
-    return [[t.t() for t in y] for y in zip(*x)]
+def _flatten(x: List[List[torch.Tensor]], rows: int, cols: int) -> List[torch.Tensor]:
+    assert len(x) == rows
+    assert all(len(row) == cols for row in x)
+    flat_x = [elem for row in x for elem in row]
+    assert len(flat_x) == rows * cols
+    return flat_x
 
 
-class _TiledMatmul(torch.autograd.Function):
-    @staticmethod
-    def forward(ctx, ab_tree_spec, *ab_tree_values):
-        ctx.ab_tree_spec = ab_tree_spec
-        ctx.save_for_backward(*ab_tree_values)
-        a, b = tree_unflatten(list(ab_tree_values), ab_tree_spec)
+def _unflatten(
+    flat_x: List[torch.Tensor], rows: int, cols: int
+) -> List[List[torch.Tensor]]:
+    assert len(flat_x) == cols * rows
+    x = [
+        flat_x[row_offset : row_offset + cols]
+        for row_offset in range(0, rows * cols, cols)
+    ]
+    assert len(x) == rows
+    assert all(len(row) == cols for row in x)
+    return x
 
-        c = tiled_matmul_fwd(a, b)
 
-        c_tree_values, c_tree_spec = tree_flatten(c)
-        ctx.c_tree_spec = c_tree_spec
-        return (c_tree_spec,) + tuple(c_tree_values)
+def _flattened_transpose(
+    flat_x: List[torch.Tensor], rows: int, cols: int
+) -> List[torch.Tensor]:
+    x = _unflatten(flat_x, rows, cols)
+    transposed_x = [[elem.t() for elem in col] for col in zip(*x)]
+    flat_transposed_x = _flatten(transposed_x, cols, rows)
+    return flat_transposed_x
 
-    @staticmethod
-    def backward(ctx, _none, *grad_c_tree_values):
-        a, b = tree_unflatten(list(ctx.saved_tensors), ctx.ab_tree_spec)
-        grad_c = tree_unflatten(list(grad_c_tree_values), ctx.c_tree_spec)
 
-        grad_a = tiled_matmul_fwd(grad_c, _transpose(b))
-        grad_b = tiled_matmul_fwd(_transpose(a), grad_c)
+# PyTorch (custom_op and torch.compile, but also the dispatcher in general)
+# have a hard time with Tensor[][] args. Thus we flatten them into Tensor[] to
+# pass them into and out of this operator.
+# See: https://github.com/pytorch/pytorch/issues/113022
+@torch.library.custom_op(
+    "xformers_python::tiled_matmul_fwd",
+    mutates_args=(),
+    device_types="cuda",
+)
+def tiled_matmul_fwd(
+    flat_a: List[torch.Tensor],
+    flat_b: List[torch.Tensor],
+    ms: List[int],
+    ns: List[int],
+    ks: List[int],
+) -> List[torch.Tensor]:
+    a = _unflatten(flat_a, len(ms), len(ks))
+    b = _unflatten(flat_b, len(ks), len(ns))
 
-        grad_ab_tree_values, grad_ab_tree_spec = tree_flatten((grad_a, grad_b))
-        return (None,) + tuple(grad_ab_tree_values)
+    c = [[a[0][0].new_empty((m, n)) for n in ns] for m in ms]
+    tiled_matmul_out(a, b, out=c)
+
+    return _flatten(c, len(ms), len(ns))
+
+
+@torch.library.register_fake("xformers_python::tiled_matmul_fwd")
+def tiled_matmul_fwd_fake(
+    flat_a: List[torch.Tensor],
+    flat_b: List[torch.Tensor],
+    ms: List[int],
+    ns: List[int],
+    ks: List[int],
+) -> List[torch.Tensor]:
+    c = [[flat_a[0][0].new_empty((m, n)) for n in ns] for m in ms]
+    return _flatten(c, len(ms), len(ks))
+
+
+def tiled_matmul_setup_context(ctx, inputs, output):
+    flat_a, flat_b, ctx.ms, ctx.ns, ctx.ks = inputs
+    ctx.save_for_backward(*flat_a, *flat_b)
+
+
+def tiled_matmul_bwd(ctx, flat_grad_c):
+    assert len(ctx.saved_tensors) == len(ctx.ms) * len(ctx.ks) + len(ctx.ks) * len(
+        ctx.ns
+    )
+    flat_a = ctx.saved_tensors[: len(ctx.ms) * len(ctx.ks)]
+    flat_b = ctx.saved_tensors[-len(ctx.ks) * len(ctx.ns) :]
+
+    flat_transposed_a = _flattened_transpose(flat_a, len(ctx.ms), len(ctx.ks))
+    flat_transposed_b = _flattened_transpose(flat_b, len(ctx.ks), len(ctx.ns))
+
+    flat_grad_a = tiled_matmul_fwd(
+        flat_grad_c, flat_transposed_b, ctx.ms, ctx.ks, ctx.ns
+    )
+    flat_grad_b = tiled_matmul_fwd(
+        flat_transposed_a, flat_grad_c, ctx.ks, ctx.ns, ctx.ms
+    )
+
+    return flat_grad_a, flat_grad_b, None, None, None
+
+
+torch.library.register_autograd(
+    "xformers_python::tiled_matmul_fwd",
+    tiled_matmul_bwd,
+    setup_context=tiled_matmul_setup_context,
+)
 
 
 def tiled_matmul(
@@ -235,8 +308,11 @@ def tiled_matmul(
     This operator is differentiable.
 
     """
-    ab_tree_values, ab_tree_spec = tree_flatten((a, b))
-    c_tree_spec, *c_tree_values = _TiledMatmul.apply(ab_tree_spec, *ab_tree_values)
-    c = tree_unflatten(list(c_tree_values), c_tree_spec)
-
+    # Inputs are checked inside the op, but we check them as well to make sure
+    # that they are "regular" and can be flattened.
+    ms, ns, ks = check_inputs(a, b)
+    flat_a = _flatten(a, len(ms), len(ks))
+    flat_b = _flatten(b, len(ks), len(ns))
+    flat_c = tiled_matmul_fwd(flat_a, flat_b, ms, ns, ks)
+    c = _unflatten(flat_c, len(ms), len(ns))
     return c